Renaming config file to project
