+++ lib/cupper/project.rb
